add SignPackagings component
